+++ util.go
@@ -11,11 +11,11 @@ const (
 )
 
 // murmurhash3 64-bit finalizer
-func hash64(x uintptr) uint32 {
+func hash64(x uintptr) uint64 {
 	x = ((x >> 33) ^ x) * 0xff51afd7ed558ccd
 	x = ((x >> 33) ^ x) * 0xc4ceb9fe1a85ec53
 	x = (x >> 33) ^ x
-	return uint32(x)
+	return uint64(x)
 }
 
 // FNV-1a 32-bit hash
